dev: cambiando double por BigDecimal para mas precision

+++ src/main/java/com/atorres/nttdata/productomicroservice/model/dao/CreditDao.java
@@ -5,12 +5,14 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @Document("credit")
 public class CreditDao {
     @Id
     private String id;
-    private Double balance;
-    private Double debt;
+    private BigDecimal balance;
+    private BigDecimal debt;
 }
