+++ src/main/java/com/atorres/nttdata/productomicroservice/service/creditstrategy/BussinesCreditStrategy.java
@@ -10,7 +10,7 @@ public class BussinesCreditStrategy implements CreditStrategy{
     @Override
     public Mono<Boolean> verifyCredit(Flux<CreditDao> listCredit) {
         return listCredit
-                .all(creditDao ->  creditDao.getBalance() <=10000)
+                .all(creditDao ->  creditDao.getBalance().doubleValue() <=10000)
                 .flatMap(  exist -> exist ? Mono.just(Boolean.TRUE):Mono.error(new CustomException(HttpStatus.BAD_REQUEST,"El credito no debe pasar 10 000")));
     }
 }
